+++ actix-utils/src/lib.rs
@@ -6,6 +6,7 @@ pub mod either;
 pub mod framed;
 pub mod inflight;
 pub mod keepalive;
+pub mod order;
 pub mod stream;
 pub mod time;
 pub mod timeout;